requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
cycler==0.10.0
decorator==5.0.9
future==0.18.2
geocoder==1.38.1
idna==2.10
kafka-python==2.0.2
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
oauthlib==3.1.0
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
ratelim==0.1.6
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
tweepy==3.10.0
urllib3==1.26.4
wordcloud==1.8.1
